Move ShortDigest out of display package
