--- cmd/oras/internal/display/status/convert.go
@@ -1,32 +1,0 @@
-/*
-Copyright The ORAS Authors.
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package status
-
-import (
-	"github.com/opencontainers/go-digest"
-	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
-)
-
-// ShortDigest converts the digest of the descriptor to a short form for displaying.
-func ShortDigest(desc ocispec.Descriptor) (digestString string) {
-	digestString = desc.Digest.String()
-	if err := desc.Digest.Validate(); err == nil {
-		if algo := desc.Digest.Algorithm(); algo == digest.SHA256 {
-			digestString = desc.Digest.Encoded()[:12]
-		}
-	}
-	return digestString
-}
